refactor(stats): separate Stats into StatsContent
